Re-Organize demuxing
