+++ src/demux/audio/aacdemuxer.ts
@@ -3,10 +3,10 @@
  */
 import BaseAudioDemuxer from './base-audio-demuxer';
 import * as ADTS from './adts';
-import { logger } from '../utils/logger';
-import * as ID3 from '../demux/id3';
-import type { HlsEventEmitter } from '../events';
-import type { HlsConfig } from '../config';
+import { logger } from '../../utils/logger';
+import * as ID3 from '../id3';
+import type { HlsEventEmitter } from '../../events';
+import type { HlsConfig } from '../../config';
 
 class AACDemuxer extends BaseAudioDemuxer {
   private readonly observer: HlsEventEmitter;
