+++ src/demux/audio/mpegaudio.ts
@@ -1,7 +1,7 @@
 /**
  *  MPEG parser helper
  */
-import { DemuxedAudioTrack } from '../types/demuxer';
+import { DemuxedAudioTrack } from '../../types/demuxer';
 
 let chromeVersion: number | null = null;
