+++ src/demux/transmuxer.ts
@@ -2,11 +2,11 @@ import type { HlsEventEmitter } from '../events';
 import { Events } from '../events';
 import { ErrorTypes, ErrorDetails } from '../errors';
 import Decrypter from '../crypt/decrypter';
-import AACDemuxer from '../demux/aacdemuxer';
+import AACDemuxer from './audio/aacdemuxer';
 import MP4Demuxer from '../demux/mp4demuxer';
 import TSDemuxer, { TypeSupported } from '../demux/tsdemuxer';
-import MP3Demuxer from '../demux/mp3demuxer';
-import { AC3Demuxer } from './ac3-demuxer';
+import MP3Demuxer from './audio/mp3demuxer';
+import { AC3Demuxer } from './audio/ac3-demuxer';
 import MP4Remuxer from '../remux/mp4-remuxer';
 import PassThroughRemuxer from '../remux/passthrough-remuxer';
 import { logger } from '../utils/logger';
