+++ src/demux/tsdemuxer.ts
@@ -9,17 +9,13 @@
  * upon discontinuity or level switch detection, it will also notifies the remuxer so that it can reset its state.
  */
 
-import * as ADTS from './adts';
-import * as MpegAudio from './mpegaudio';
-import * as AC3 from './ac3-demuxer';
-import ExpGolomb from './exp-golomb';
+import * as ADTS from './audio/adts';
+import * as MpegAudio from './audio/mpegaudio';
+import * as AC3 from './audio/ac3-demuxer';
+import AvcVideoParser from './video/avc-video-parser';
 import SampleAesDecrypter from './sample-aes';
 import { Events } from '../events';
-import {
-  appendUint8Array,
-  parseSEIMessageFromNALu,
-  RemuxerTrackIdConfig,
-} from '../utils/mp4-tools';
+import { appendUint8Array, RemuxerTrackIdConfig } from '../utils/mp4-tools';
 import { logger } from '../utils/logger';
 import { ErrorTypes, ErrorDetails } from '../errors';
 import type { HlsConfig } from '../config';
@@ -36,21 +32,21 @@ import {
   ElementaryStreamData,
   KeyData,
   MetadataSchema,
-  VideoSampleUnit,
 } from '../types/demuxer';
 import { AudioFrame } from '../types/demuxer';
 
-type ParsedTimestamp = {
+export type ParsedTimestamp = {
   pts?: number;
   dts?: number;
 };
 
-type PES = ParsedTimestamp & {
+export type PES = ParsedTimestamp & {
   data: Uint8Array;
   len: number;
 };
 
-type ParsedVideoSample = ParsedTimestamp & Omit<VideoSample, 'pts' | 'dts'>;
+export type ParsedVideoSample = ParsedTimestamp &
+  Omit<VideoSample, 'pts' | 'dts'>;
 
 export interface TypeSupported {
   mpeg: boolean;
@@ -77,8 +73,8 @@ class TSDemuxer implements Demuxer {
   private _id3Track?: DemuxedMetadataTrack;
   private _txtTrack?: DemuxedUserdataTrack;
   private aacOverFlow: AudioFrame | null = null;
-  private VideoSample: ParsedVideoSample | null = null;
   private remainderData: Uint8Array | null = null;
+  private videoParser: AvcVideoParser;
 
   constructor(
     observer: HlsEventEmitter,
@@ -88,6 +84,7 @@ class TSDemuxer implements Demuxer {
     this.observer = observer;
     this.config = config;
     this.typeSupported = typeSupported;
+    this.videoParser = new AvcVideoParser();
   }
 
   static probe(data: Uint8Array) {
@@ -193,7 +190,6 @@ class TSDemuxer implements Demuxer {
 
     // flush any partial content
     this.aacOverFlow = null;
-    this.VideoSample = null;
     this.remainderData = null;
     this.audioCodec = audioCodec;
     this.videoCodec = videoCodec;
@@ -214,7 +210,6 @@ class TSDemuxer implements Demuxer {
       _id3Track.pesData = null;
     }
     this.aacOverFlow = null;
-    this.VideoSample = null;
     this.remainderData = null;
   }
 
@@ -295,7 +290,13 @@ class TSDemuxer implements Demuxer {
           case videoPid:
             if (stt) {
               if (videoData && (pes = parsePES(videoData))) {
-                this.parseAVCPES(videoTrack, textTrack, pes, false);
+                this.videoParser.parseAVCPES(
+                  videoTrack,
+                  textTrack,
+                  pes,
+                  false,
+                  this._duration
+                );
               }
 
               videoData = { data: [], size: 0 };
@@ -371,12 +372,13 @@ class TSDemuxer implements Demuxer {
             videoPid = parsedPIDs.videoPid;
             if (videoPid > 0) {
               videoTrack.pid = videoPid;
+              videoTrack.segmentCodec = parsedPIDs.segmentVideoCodec;
             }
 
             audioPid = parsedPIDs.audioPid;
             if (audioPid > 0) {
               audioTrack.pid = audioPid;
-              audioTrack.segmentCodec = parsedPIDs.segmentCodec;
+              audioTrack.segmentCodec = parsedPIDs.segmentAudioCodec;
             }
             id3Pid = parsedPIDs.id3Pid;
             if (id3Pid > 0) {
@@ -466,11 +468,12 @@ class TSDemuxer implements Demuxer {
     // try to parse last PES packets
     let pes: PES | null;
     if (videoData && (pes = parsePES(videoData))) {
-      this.parseAVCPES(
+      this.videoParser.parseAVCPES(
         videoTrack as DemuxedVideoTrack,
         textTrack as DemuxedUserdataTrack,
         pes,
-        true
+        true,
+        this._duration
       );
       videoTrack.pesData = null;
     } else {
@@ -558,320 +561,6 @@ class TSDemuxer implements Demuxer {
 
   public destroy() {
     this._duration = 0;
-  }
-
-  private parseAVCPES(
-    track: DemuxedVideoTrack,
-    textTrack: DemuxedUserdataTrack,
-    pes: PES,
-    last: boolean
-  ) {
-    const units = this.parseAVCNALu(track, pes.data);
-    const debug = false;
-    let VideoSample = this.VideoSample;
-    let push: boolean;
-    let spsfound = false;
-    // free pes.data to save up some memory
-    (pes as any).data = null;
-
-    // if new NAL units found and last sample still there, let's push ...
-    // this helps parsing streams with missing AUD (only do this if AUD never found)
-    if (VideoSample && units.length && !track.audFound) {
-      pushAccessUnit(VideoSample, track);
-      VideoSample = this.VideoSample = createVideoSample(
-        false,
-        pes.pts,
-        pes.dts,
-        ''
-      );
-    }
-
-    units.forEach((unit) => {
-      switch (unit.type) {
-        // NDR
-        case 1: {
-          push = true;
-          if (!VideoSample) {
-            VideoSample = this.VideoSample = createVideoSample(
-              true,
-              pes.pts,
-              pes.dts,
-              ''
-            );
-          }
-
-          if (debug) {
-            VideoSample.debug += 'NDR ';
-          }
-
-          VideoSample.frame = true;
-          const data = unit.data;
-          // only check slice type to detect KF in case SPS found in same packet (any keyframe is preceded by SPS ...)
-          if (spsfound && data.length > 4) {
-            // retrieve slice type by parsing beginning of NAL unit (follow H264 spec, slice_header definition) to detect keyframe embedded in NDR
-            const sliceType = new ExpGolomb(data).readSliceType();
-            // 2 : I slice, 4 : SI slice, 7 : I slice, 9: SI slice
-            // SI slice : A slice that is coded using intra prediction only and using quantisation of the prediction samples.
-            // An SI slice can be coded such that its decoded samples can be constructed identically to an SP slice.
-            // I slice: A slice that is not an SI slice that is decoded using intra prediction only.
-            // if (sliceType === 2 || sliceType === 7) {
-            if (
-              sliceType === 2 ||
-              sliceType === 4 ||
-              sliceType === 7 ||
-              sliceType === 9
-            ) {
-              VideoSample.key = true;
-            }
-          }
-          break;
-          // IDR
-        }
-        case 5:
-          push = true;
-          // handle PES not starting with AUD
-          if (!VideoSample) {
-            VideoSample = this.VideoSample = createVideoSample(
-              true,
-              pes.pts,
-              pes.dts,
-              ''
-            );
-          }
-
-          if (debug) {
-            VideoSample.debug += 'IDR ';
-          }
-
-          VideoSample.key = true;
-          VideoSample.frame = true;
-          break;
-        // SEI
-        case 6: {
-          push = true;
-          if (debug && VideoSample) {
-            VideoSample.debug += 'SEI ';
-          }
-          parseSEIMessageFromNALu(
-            unit.data,
-            1,
-            pes.pts as number,
-            textTrack.samples
-          );
-          break;
-          // SPS
-        }
-        case 7:
-          push = true;
-          spsfound = true;
-          if (debug && VideoSample) {
-            VideoSample.debug += 'SPS ';
-          }
-
-          if (!track.sps) {
-            const sps = unit.data;
-            const expGolombDecoder = new ExpGolomb(sps);
-            const config = expGolombDecoder.readSPS();
-            track.width = config.width;
-            track.height = config.height;
-            track.pixelRatio = config.pixelRatio;
-            track.sps = [sps];
-            track.duration = this._duration;
-            const codecarray = sps.subarray(1, 4);
-            let codecstring = 'avc1.';
-            for (let i = 0; i < 3; i++) {
-              let h = codecarray[i].toString(16);
-              if (h.length < 2) {
-                h = '0' + h;
-              }
-
-              codecstring += h;
-            }
-            track.codec = codecstring;
-          }
-          break;
-        // PPS
-        case 8:
-          push = true;
-          if (debug && VideoSample) {
-            VideoSample.debug += 'PPS ';
-          }
-
-          if (!track.pps) {
-            track.pps = [unit.data];
-          }
-
-          break;
-        // AUD
-        case 9:
-          push = false;
-          track.audFound = true;
-          if (VideoSample) {
-            pushAccessUnit(VideoSample, track);
-          }
-
-          VideoSample = this.VideoSample = createVideoSample(
-            false,
-            pes.pts,
-            pes.dts,
-            debug ? 'AUD ' : ''
-          );
-          break;
-        // Filler Data
-        case 12:
-          push = true;
-          break;
-        default:
-          push = false;
-          if (VideoSample) {
-            VideoSample.debug += 'unknown NAL ' + unit.type + ' ';
-          }
-
-          break;
-      }
-      if (VideoSample && push) {
-        const units = VideoSample.units;
-        units.push(unit);
-      }
-    });
-    // if last PES packet, push samples
-    if (last && VideoSample) {
-      pushAccessUnit(VideoSample, track);
-      this.VideoSample = null;
-    }
-  }
-
-  private getLastNalUnit(samples: VideoSample[]): VideoSampleUnit | undefined {
-    let VideoSample = this.VideoSample;
-    let lastUnit: VideoSampleUnit | undefined;
-    // try to fallback to previous sample if current one is empty
-    if (!VideoSample || VideoSample.units.length === 0) {
-      VideoSample = samples[samples.length - 1];
-    }
-    if (VideoSample?.units) {
-      const units = VideoSample.units;
-      lastUnit = units[units.length - 1];
-    }
-    return lastUnit;
-  }
-
-  private parseAVCNALu(
-    track: DemuxedVideoTrack,
-    array: Uint8Array
-  ): Array<{
-    data: Uint8Array;
-    type: number;
-    state?: number;
-  }> {
-    const len = array.byteLength;
-    let state = track.naluState || 0;
-    const lastState = state;
-    const units: VideoSampleUnit[] = [];
-    let i = 0;
-    let value: number;
-    let overflow: number;
-    let unitType: number;
-    let lastUnitStart = -1;
-    let lastUnitType: number = 0;
-    // logger.log('PES:' + Hex.hexDump(array));
-
-    if (state === -1) {
-      // special use case where we found 3 or 4-byte start codes exactly at the end of previous PES packet
-      lastUnitStart = 0;
-      // NALu type is value read from offset 0
-      lastUnitType = array[0] & 0x1f;
-      state = 0;
-      i = 1;
-    }
-
-    while (i < len) {
-      value = array[i++];
-      // optimization. state 0 and 1 are the predominant case. let's handle them outside of the switch/case
-      if (!state) {
-        state = value ? 0 : 1;
-        continue;
-      }
-      if (state === 1) {
-        state = value ? 0 : 2;
-        continue;
-      }
-      // here we have state either equal to 2 or 3
-      if (!value) {
-        state = 3;
-      } else if (value === 1) {
-        overflow = i - state - 1;
-        if (lastUnitStart >= 0) {
-          const unit: VideoSampleUnit = {
-            data: array.subarray(lastUnitStart, overflow),
-            type: lastUnitType,
-          };
-          // logger.log('pushing NALU, type/size:' + unit.type + '/' + unit.data.byteLength);
-          units.push(unit);
-        } else {
-          // lastUnitStart is undefined => this is the first start code found in this PES packet
-          // first check if start code delimiter is overlapping between 2 PES packets,
-          // ie it started in last packet (lastState not zero)
-          // and ended at the beginning of this PES packet (i <= 4 - lastState)
-          const lastUnit = this.getLastNalUnit(track.samples);
-          if (lastUnit) {
-            if (lastState && i <= 4 - lastState) {
-              // start delimiter overlapping between PES packets
-              // strip start delimiter bytes from the end of last NAL unit
-              // check if lastUnit had a state different from zero
-              if (lastUnit.state) {
-                // strip last bytes
-                lastUnit.data = lastUnit.data.subarray(
-                  0,
-                  lastUnit.data.byteLength - lastState
-                );
-              }
-            }
-            // If NAL units are not starting right at the beginning of the PES packet, push preceding data into previous NAL unit.
-
-            if (overflow > 0) {
-              // logger.log('first NALU found with overflow:' + overflow);
-              lastUnit.data = appendUint8Array(
-                lastUnit.data,
-                array.subarray(0, overflow)
-              );
-              lastUnit.state = 0;
-            }
-          }
-        }
-        // check if we can read unit type
-        if (i < len) {
-          unitType = array[i] & 0x1f;
-          // logger.log('find NALU @ offset:' + i + ',type:' + unitType);
-          lastUnitStart = i;
-          lastUnitType = unitType;
-          state = 0;
-        } else {
-          // not enough byte to read unit type. let's read it on next PES parsing
-          state = -1;
-        }
-      } else {
-        state = 0;
-      }
-    }
-    if (lastUnitStart >= 0 && state >= 0) {
-      const unit: VideoSampleUnit = {
-        data: array.subarray(lastUnitStart, len),
-        type: lastUnitType,
-        state: state,
-      };
-      units.push(unit);
-      // logger.log('pushing NALU, type/size/state:' + unit.type + '/' + unit.data.byteLength + '/' + state);
-    }
-    // no NALu found
-    if (units.length === 0) {
-      // append pes.data to previous NAL unit
-      const lastUnit = this.getLastNalUnit(track.samples);
-      if (lastUnit) {
-        lastUnit.data = appendUint8Array(lastUnit.data, array);
-      }
-    }
-    track.naluState = state;
-    return units;
   }
 
   private parseAACPES(track: DemuxedAudioTrack, pes: PES) {
@@ -1036,23 +725,6 @@ class TSDemuxer implements Demuxer {
   }
 }
 
-function createVideoSample(
-  key: boolean,
-  pts: number | undefined,
-  dts: number | undefined,
-  debug: string
-): ParsedVideoSample {
-  return {
-    key,
-    frame: false,
-    pts,
-    dts,
-    units: [],
-    debug,
-    length: 0,
-  };
-}
-
 function parsePID(data: Uint8Array, offset: number): number {
   // pid is a 13-bit field starting at the last bit of TS[1]
   return ((data[offset + 1] & 0x1f) << 8) + data[offset + 2];
@@ -1073,7 +745,8 @@ function parsePMT(
     audioPid: -1,
     videoPid: -1,
     id3Pid: -1,
-    segmentCodec: 'aac',
+    segmentVideoCodec: 'avc',
+    segmentAudioCodec: 'aac',
   };
   const sectionLength = ((data[offset + 1] & 0x0f) << 8) | data[offset + 2];
   const tableEnd = offset + 3 + sectionLength - 4;
@@ -1119,6 +792,7 @@ function parsePMT(
         // logger.log('AVC PID:'  + pid);
         if (result.videoPid === -1) {
           result.videoPid = pid;
+          result.segmentVideoCodec = 'avc';
         }
 
         break;
@@ -1132,7 +806,7 @@ function parsePMT(
           logger.log('MPEG audio found, not supported in this browser');
         } else if (result.audioPid === -1) {
           result.audioPid = pid;
-          result.segmentCodec = 'mp3';
+          result.segmentAudioCodec = 'mp3';
         }
         break;
 
@@ -1148,7 +822,7 @@ function parsePMT(
             logger.log('AC-3 audio found, not supported in this browser');
           } else if (result.audioPid === -1) {
             result.audioPid = pid;
-            result.segmentCodec = 'ac3';
+            result.segmentAudioCodec = 'ac3';
           }
         } else {
           logger.warn('AC-3 in M2TS support not included in build');
@@ -1276,34 +950,6 @@ function parsePES(stream: ElementaryStreamData): PES | null {
     return { data: pesData, pts: pesPts, dts: pesDts, len: pesLen };
   }
   return null;
-}
-
-function pushAccessUnit(
-  VideoSample: ParsedVideoSample,
-  videoTrack: DemuxedVideoTrack
-) {
-  if (VideoSample.units.length && VideoSample.frame) {
-    // if sample does not have PTS/DTS, patch with last sample PTS/DTS
-    if (VideoSample.pts === undefined) {
-      const samples = videoTrack.samples;
-      const nbSamples = samples.length;
-      if (nbSamples) {
-        const lastSample = samples[nbSamples - 1];
-        VideoSample.pts = lastSample.pts;
-        VideoSample.dts = lastSample.dts;
-      } else {
-        // dropping samples, no timestamp found
-        videoTrack.dropped++;
-        return;
-      }
-    }
-    videoTrack.samples.push(VideoSample as VideoSample);
-  }
-  if (VideoSample.debug.length) {
-    logger.log(
-      VideoSample.pts + '/' + VideoSample.dts + ':' + VideoSample.debug
-    );
-  }
 }
 
 export default TSDemuxer;
