+++ src/types/demuxer.ts
@@ -75,6 +75,7 @@ export interface DemuxedVideoTrackBase extends DemuxedTrack {
   pps?: Uint8Array[];
   sps?: Uint8Array[];
   naluState?: number;
+  segmentCodec?: string;
   manifestCodec?: string;
   samples: VideoSample[] | Uint8Array;
 }
